// Frontend/src/app/models/documento.ts
export interface documento {
    Doc_ID?: number,
    Nombre?: string;
    Descripcion?: string;
    Proy_ID?: number;
    URL?: string;
}

export interface docc {
    Nombre?: string;
    Descripcion?: string;
    Proy_ID?: number;
    URL?: string;
} 


export interface docup {
    Nombre?: string;
    Descripcion?: string;
    URL?: string;
    Doc_ID?: number;
} 